Rename semi_stage and move semisl model config to otx
