+++ mpa/det/semisl/trainer.py
@@ -23,7 +23,7 @@ from mmdet.utils import collect_env
 from mpa.registry import STAGES
 from mpa.modules.utils.task_adapt import extract_anchor_ratio
 from mpa.utils.logger import get_logger
-from mpa.det.semisl.semisl_stage import SemiSLDetectionStage
+from mpa.det.semisl.stage import SemiSLDetectionStage
 
 logger = get_logger()
 
@@ -40,6 +40,7 @@ class SemiSLDetectionTrainer(SemiSLDetectionStage):
         - Environment setup
         - Run training via MMDetection -> MMCV
         """
+        breakpoint()
         self._init_logger()
         mode = kwargs.get('mode', 'train')
         if mode not in self.mode:
